+++ packages/bpk-component-rtl-toggle/src/BpkRtlToggle-test.js
@@ -33,4 +33,9 @@ describe('BpkRtlToggle', () => {
     tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
+
+  it('should render correctly with the className prop', () => {
+    const tree = renderer.create(<BpkRtlToggle className="foo" />).toJSON();
+    expect(tree).toMatchSnapshot();
+  });
 });